Split functionality into files
